refactor!: remove unnecessary error enum and result type aliases (#7875)
