+++ .changes/tauri-extranous-result-types-removal.md
@@ -1,0 +1,5 @@
+---
+'tauri': 'major:breaking'
+---
+
+Removed `tauri::path::Error` and added its variants to `tauri::Error`.
